+++ lib/src/bluez_media_client.dart
@@ -35,36 +35,18 @@ export 'ffi/types.dart'
 class BluezMediaPlayerRegistrationConfig {
   final String adapterPath;
   final String playerPath;
-  final String identity;
   final String name;
   final String type;
   final String subtype;
-  final String status;
-  final int positionMs;
-  final bool canGoNext;
-  final bool canGoPrevious;
-  final bool canPlay;
-  final bool canPause;
-  final bool canSeek;
-  final bool canControl;
   final bool browsable;
   final bool searchable;
 
   const BluezMediaPlayerRegistrationConfig({
     required this.adapterPath,
     required this.playerPath,
-    this.identity = 'bluez_media_native',
     this.name = 'bluez_media_native',
     this.type = 'audio',
     this.subtype = '',
-    this.status = 'stopped',
-    this.positionMs = 0,
-    this.canGoNext = false,
-    this.canGoPrevious = false,
-    this.canPlay = true,
-    this.canPause = true,
-    this.canSeek = false,
-    this.canControl = true,
     this.browsable = false,
     this.searchable = false,
   });
@@ -211,18 +193,9 @@ class BluezMediaClient {
       registration.ref
         ..adapter_path = nativeString(config.adapterPath)
         ..player_path = nativeString(config.playerPath)
-        ..identity = nativeString(config.identity)
         ..name = nativeString(config.name)
         ..type = nativeString(config.type)
         ..subtype = nativeString(config.subtype)
-        ..status = nativeString(config.status)
-        ..position_ms = config.positionMs
-        ..can_go_next = config.canGoNext ? 1 : 0
-        ..can_go_previous = config.canGoPrevious ? 1 : 0
-        ..can_play = config.canPlay ? 1 : 0
-        ..can_pause = config.canPause ? 1 : 0
-        ..can_seek = config.canSeek ? 1 : 0
-        ..can_control = config.canControl ? 1 : 0
         ..browsable = config.browsable ? 1 : 0
         ..searchable = config.searchable ? 1 : 0;
 
@@ -766,12 +739,6 @@ class BluezMediaClient {
     }
   }
 
-  List<BluezMediaTransport> refreshTransports() {
-    return getManagedObjects().transports
-        .map(transport)
-        .toList(growable: false);
-  }
-
   void closeFileDescriptor(int fd) {
     final result = _bindings.bluez_media_close_fd(fd);
     _checkResult(result, 'close acquired media transport file descriptor');
@@ -904,33 +871,6 @@ class BluezMediaClient {
   }
 }
 
-/// A very short-lived native function.
-///
-/// For very short-lived functions, it is fine to call them on the main isolate.
-/// They will block the Dart execution while running the native function, so
-/// only do this for native functions which are guaranteed to be short-lived.
-int sum(int a, int b) => _bindings.sum(a, b);
-
-/// A longer lived native function, which occupies the thread calling it.
-///
-/// Do not call these kind of native functions in the main isolate. They will
-/// block Dart execution. This will cause dropped frames in Flutter applications.
-/// Instead, call these native functions on a separate isolate.
-///
-/// Modify this to suit your own use case. Example use cases:
-///
-/// 1. Reuse a single isolate for various different kinds of requests.
-/// 2. Use multiple helper isolates for parallel execution.
-Future<int> sumAsync(int a, int b) async {
-  final SendPort helperIsolateSendPort = await _helperIsolateSendPort;
-  final int requestId = _nextSumRequestId++;
-  final _SumRequest request = _SumRequest(requestId, a, b);
-  final Completer<int> completer = Completer<int>();
-  _sumRequests[requestId] = completer;
-  helperIsolateSendPort.send(request);
-  return completer.future;
-}
-
 /// The dynamic library in which the symbols for [BluezMediaNativeBindings] can be found.
 final DynamicLibrary _dylib = loadBluezMediaNative();
 
@@ -943,80 +883,3 @@ void _initializeNativeApi() {
   _bindings.bluez_media_init(NativeApi.initializeApiDLData);
   _nativeApiInitialized = true;
 }
-
-/// A request to compute `sum`.
-///
-/// Typically sent from one isolate to another.
-class _SumRequest {
-  final int id;
-  final int a;
-  final int b;
-
-  const _SumRequest(this.id, this.a, this.b);
-}
-
-/// A response with the result of `sum`.
-///
-/// Typically sent from one isolate to another.
-class _SumResponse {
-  final int id;
-  final int result;
-
-  const _SumResponse(this.id, this.result);
-}
-
-/// Counter to identify [_SumRequest]s and [_SumResponse]s.
-int _nextSumRequestId = 0;
-
-/// Mapping from [_SumRequest] `id`s to the completers corresponding to the correct future of the pending request.
-final Map<int, Completer<int>> _sumRequests = <int, Completer<int>>{};
-
-/// The SendPort belonging to the helper isolate.
-Future<SendPort> _helperIsolateSendPort = () async {
-  // The helper isolate is going to send us back a SendPort, which we want to
-  // wait for.
-  final Completer<SendPort> completer = Completer<SendPort>();
-
-  // Receive port on the main isolate to receive messages from the helper.
-  // We receive two types of messages:
-  // 1. A port to send messages on.
-  // 2. Responses to requests we sent.
-  final ReceivePort receivePort = ReceivePort()
-    ..listen((dynamic data) {
-      if (data is SendPort) {
-        // The helper isolate sent us the port on which we can sent it requests.
-        completer.complete(data);
-        return;
-      }
-      if (data is _SumResponse) {
-        // The helper isolate sent us a response to a request we sent.
-        final Completer<int> completer = _sumRequests[data.id]!;
-        _sumRequests.remove(data.id);
-        completer.complete(data.result);
-        return;
-      }
-      throw UnsupportedError('Unsupported message type: ${data.runtimeType}');
-    });
-
-  // Start the helper isolate.
-  await Isolate.spawn((SendPort sendPort) async {
-    final ReceivePort helperReceivePort = ReceivePort()
-      ..listen((dynamic data) {
-        // On the helper isolate listen to requests and respond to them.
-        if (data is _SumRequest) {
-          final int result = _bindings.sum_long_running(data.a, data.b);
-          final _SumResponse response = _SumResponse(data.id, result);
-          sendPort.send(response);
-          return;
-        }
-        throw UnsupportedError('Unsupported message type: ${data.runtimeType}');
-      });
-
-    // Send the port to the main isolate on which we can receive requests.
-    sendPort.send(helperReceivePort.sendPort);
-  }, receivePort.sendPort);
-
-  // Wait until the helper isolate has sent us back the SendPort on which we
-  // can start sending requests.
-  return completer.future;
-}();